Adding more funtionality to XSD and Service

+++ src/main/java/dom/soapexample/model/Book.java
@@ -24,5 +24,5 @@ public class Book implements Serializable {
     private String title;
     private String author;
     private BigDecimal price;
-    private BigInteger stock;
+    private int stock;
 }
